Bugfix: Use DOM loaded event for TOC
The document "DOMContentLoaded" event will fire as soon as the DOM is
ready to be manipulated by the TOC building script.

The window.onload function should not be used for this because it waits
until the window is rendered in the browser to run the script, causing a
flash of partially rendered content.