repositories: authenticated installation instructions for ubuntu
These instructions were used for the Debian repositories, but we
neglected to also use them for Ubuntu.

Also, while at it, use a separate file in /etc/apt/sources.list.d
instead of appending to the system-wide /etc/apt/sources.list, and use
grep+cut instead of lsb_release(1) which may not be installed on
minimal Ubuntu installations (the xenial Docker image lacks it, for
example).