Include a basic CSS reset and manage the content directly in the main element
This is a bit of a harsher change. By including a CSS reset we mess up the parts
of the layout that depend on the default browser styles, like p and ul tags.

To compensate for this we add those styles to our main content element, which
we wrap around all of our content in the default template file. Since we will
use the <main> tag for our content spacing purpses moving forward we can deprecate the
index files, but we'll leave it in the stylesheet for backwards compatability.