Update the refcard
I've made a couple of changes to the refcard HTML. FOr starters, I've closed all the HTML tags that didn't have a closing tag.
I've also replaced the keyboard font with text versions of the keys, so we can drop that dependency. I've updated a couple of font-size, colours and spacings to make the refcard easier to scan.
Lastly, I added better print media queries so we can print the entire refcard on a single A4 Portrait paper.